[VM] Improve quality of codegen for crypto operations
After the Smi size reduction we are more dependent on good quality
code when using the uint32 truncating unboxed data type in the
optimizing compiler. This change lets us emit instructions with
embedded immediates for that code, instead of always loading the
immediates into a register and doing 3-register operations.

R=vegorov@google.com

Change-Id: I64e3013445d165c322bf0c3d9ee23cc3314d778d
Reviewed-on: https://dart-review.googlesource.com/50401
Reviewed-by: Vyacheslav Egorov <vegorov@google.com>